Restructure package responsive
